[WFLY-2547] Forward port naming 1.4 xsd from EAP 6.2
