+++ naming/src/main/java/org/jboss/as/naming/subsystem/NamingSubsystem20Parser.java
@@ -22,306 +22,14 @@
 
 package org.jboss.as.naming.subsystem;
 
-import java.util.EnumSet;
-import java.util.List;
-
-import javax.xml.stream.XMLStreamConstants;
-import javax.xml.stream.XMLStreamException;
-
-import org.jboss.as.controller.PathAddress;
-import org.jboss.as.controller.operations.common.Util;
-import org.jboss.dmr.ModelNode;
-import org.jboss.staxmapper.XMLElementReader;
-import org.jboss.staxmapper.XMLExtendedStreamReader;
-
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP_ADDR;
-import static org.jboss.as.controller.parsing.ParseUtils.missingRequired;
-import static org.jboss.as.controller.parsing.ParseUtils.requireAttributes;
-import static org.jboss.as.controller.parsing.ParseUtils.requireNoAttributes;
-import static org.jboss.as.controller.parsing.ParseUtils.requireNoContent;
-import static org.jboss.as.controller.parsing.ParseUtils.requireNoNamespaceAttribute;
-import static org.jboss.as.controller.parsing.ParseUtils.unexpectedAttribute;
-import static org.jboss.as.controller.parsing.ParseUtils.unexpectedElement;
-import static org.jboss.as.naming.subsystem.NamingExtension.SUBSYSTEM_PATH;
-import static org.jboss.as.naming.subsystem.NamingSubsystemModel.BINDING;
-import static org.jboss.as.naming.subsystem.NamingSubsystemModel.BINDING_TYPE;
-import static org.jboss.as.naming.subsystem.NamingSubsystemModel.EXTERNAL_CONTEXT;
-import static org.jboss.as.naming.subsystem.NamingSubsystemModel.LOOKUP;
-import static org.jboss.as.naming.subsystem.NamingSubsystemModel.OBJECT_FACTORY;
-import static org.jboss.as.naming.subsystem.NamingSubsystemModel.SIMPLE;
-
 /**
  * @author Eduardo Martins
  */
-public class NamingSubsystem20Parser implements XMLElementReader<List<ModelNode>> {
+public class NamingSubsystem20Parser extends NamingSubsystem14Parser {
 
     public static final NamingSubsystem20Parser INSTANCE = new NamingSubsystem20Parser();
 
     private NamingSubsystem20Parser() {
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void readElement(final XMLExtendedStreamReader reader, final List<ModelNode> operations) throws XMLStreamException {
-
-        PathAddress address = PathAddress.pathAddress(SUBSYSTEM_PATH);
-        final ModelNode ejb3SubsystemAddOperation = Util.createAddOperation(address);
-        operations.add(ejb3SubsystemAddOperation);
-
-        // elements
-        final EnumSet<NamingSubsystemXMLElement> encountered = EnumSet.noneOf(NamingSubsystemXMLElement.class);
-        while (reader.hasNext() && reader.nextTag() != XMLStreamConstants.END_ELEMENT) {
-            switch (NamingSubsystemNamespace.forUri(reader.getNamespaceURI())) {
-                case NAMING_2_0: {
-                    final NamingSubsystemXMLElement element = NamingSubsystemXMLElement.forName(reader.getLocalName());
-                    if (!encountered.add(element)) {
-                        throw unexpectedElement(reader);
-                    }
-                    switch (element) {
-                        case BINDINGS: {
-                            parseBindings(reader, operations, address);
-                            break;
-                        }
-                        case REMOTE_NAMING: {
-                            parseRemoteNaming(reader, operations, address);
-                            break;
-                        }
-                        default: {
-                            throw unexpectedElement(reader);
-                        }
-                    }
-                    break;
-                }
-                default: {
-                    throw unexpectedElement(reader);
-                }
-            }
-        }
-    }
-
-    private void parseRemoteNaming(final XMLExtendedStreamReader reader, final List<ModelNode> operations, PathAddress parent) throws XMLStreamException {
-        requireNoAttributes(reader);
-        requireNoContent(reader);
-        operations.add(Util.createAddOperation(parent.append(NamingSubsystemModel.SERVICE, NamingSubsystemModel.REMOTE_NAMING)));
-    }
-
-
-    private void parseBindings(final XMLExtendedStreamReader reader, final List<ModelNode> operations, PathAddress address) throws XMLStreamException {
-        // no attributes expected
-        requireNoAttributes(reader);
-        while (reader.hasNext() && reader.nextTag() != XMLStreamConstants.END_ELEMENT) {
-            switch (NamingSubsystemXMLElement.forName(reader.getLocalName())) {
-                case SIMPLE: {
-                    this.parseSimpleBinding(reader, operations, address);
-                    break;
-                }
-                case OBJECT_FACTORY: {
-                    this.parseObjectFactoryBinding(reader, operations, address);
-                    break;
-                }
-                case LOOKUP: {
-                    this.parseLookupBinding(reader, operations, address);
-                    break;
-                }
-                case EXTERNAL_CONTEXT: {
-                    this.parseExternalContext(reader, operations, address);
-                    break;
-                }
-                default: {
-                    throw unexpectedElement(reader);
-                }
-            }
-        }
-    }
-
-
-    private void parseSimpleBinding(final XMLExtendedStreamReader reader, List<ModelNode> operations, PathAddress parentAddress) throws XMLStreamException {
-        String name = null;
-        final ModelNode bindingAdd = Util.createAddOperation();
-        bindingAdd.get(BINDING_TYPE).set(SIMPLE);
-        final EnumSet<NamingSubsystemXMLAttribute> required = EnumSet.of(NamingSubsystemXMLAttribute.NAME, NamingSubsystemXMLAttribute.VALUE);
-        for (int i = 0; i < reader.getAttributeCount(); i++) {
-            requireNoNamespaceAttribute(reader, i);
-            final String value = reader.getAttributeValue(i);
-            final NamingSubsystemXMLAttribute attribute = NamingSubsystemXMLAttribute.forName(reader.getAttributeLocalName(i));
-            required.remove(attribute);
-            switch (attribute) {
-                case NAME:
-                    name = value.trim();
-                    break;
-                case VALUE:
-                    NamingBindingResourceDefinition.VALUE.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                case TYPE:
-                    NamingBindingResourceDefinition.TYPE.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                default:
-                    throw unexpectedAttribute(reader, i);
-            }
-        }
-        if (!required.isEmpty()) {
-            throw missingRequired(reader, required);
-        }
-
-        requireNoContent(reader);
-        final PathAddress address = parentAddress.append(BINDING, name);
-        bindingAdd.get(OP_ADDR).set(address.toModelNode());
-        operations.add(bindingAdd);
-    }
-
-
-    private void parseObjectFactoryBinding(final XMLExtendedStreamReader reader, List<ModelNode> operations, PathAddress parentAddress) throws XMLStreamException {
-        final int attCount = reader.getAttributeCount();
-        String name = null;
-        final ModelNode bindingAdd = Util.createAddOperation();
-        bindingAdd.get(BINDING_TYPE).set(OBJECT_FACTORY);
-        final EnumSet<NamingSubsystemXMLAttribute> required = EnumSet.of(NamingSubsystemXMLAttribute.NAME, NamingSubsystemXMLAttribute.MODULE, NamingSubsystemXMLAttribute.CLASS);
-        for (int i = 0; i < attCount; i++) {
-            requireNoNamespaceAttribute(reader, i);
-            final String value = reader.getAttributeValue(i);
-            final NamingSubsystemXMLAttribute attribute = NamingSubsystemXMLAttribute.forName(reader.getAttributeLocalName(i));
-            required.remove(attribute);
-            switch (attribute) {
-                case NAME:
-                    name = value.trim();
-                    break;
-                case MODULE:
-                    NamingBindingResourceDefinition.MODULE.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                case CLASS:
-                    NamingBindingResourceDefinition.CLASS.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                default:
-                    throw unexpectedAttribute(reader, i);
-            }
-        }
-        if (!required.isEmpty()) {
-            throw missingRequired(reader, required);
-        }
-        bindingAdd.get(OP_ADDR).set(parentAddress.append(BINDING, name).toModelNode());
-        // if present, parse the optional environment
-        while (reader.hasNext() && reader.nextTag() != XMLStreamConstants.END_ELEMENT) {
-            switch (NamingSubsystemXMLElement.forName(reader.getLocalName())) {
-                case ENVIRONMENT: {
-                    parseObjectFactoryBindingEnvironment(reader, bindingAdd);
-                    break;
-                }
-                default: {
-                    throw unexpectedElement(reader);
-                }
-            }
-        }
-
-        operations.add(bindingAdd);
-    }
-
-
-    private void parseExternalContext(final XMLExtendedStreamReader reader, List<ModelNode> operations, PathAddress parentAddress) throws XMLStreamException {
-        final int attCount = reader.getAttributeCount();
-        String name = null;
-        final ModelNode bindingAdd = Util.createAddOperation();
-        bindingAdd.get(BINDING_TYPE).set(EXTERNAL_CONTEXT);
-        final EnumSet<NamingSubsystemXMLAttribute> required = EnumSet.of(NamingSubsystemXMLAttribute.NAME, NamingSubsystemXMLAttribute.CLASS, NamingSubsystemXMLAttribute.MODULE);
-        for (int i = 0; i < attCount; i++) {
-            requireNoNamespaceAttribute(reader, i);
-            final String value = reader.getAttributeValue(i);
-            final NamingSubsystemXMLAttribute attribute = NamingSubsystemXMLAttribute.forName(reader.getAttributeLocalName(i));
-            required.remove(attribute);
-            switch (attribute) {
-                case NAME:
-                    name = value.trim();
-                    break;
-                case MODULE:
-                    NamingBindingResourceDefinition.MODULE.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                case CLASS:
-                    NamingBindingResourceDefinition.CLASS.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                case CACHE:
-                    NamingBindingResourceDefinition.CACHE.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                default:
-                    throw unexpectedAttribute(reader, i);
-            }
-        }
-        if (!required.isEmpty()) {
-            throw missingRequired(reader, required);
-        }
-        bindingAdd.get(OP_ADDR).set(parentAddress.append(BINDING, name).toModelNode());
-        // if present, parse the optional environment
-        while (reader.hasNext() && reader.nextTag() != XMLStreamConstants.END_ELEMENT) {
-            switch (NamingSubsystemXMLElement.forName(reader.getLocalName())) {
-                case ENVIRONMENT: {
-                    parseObjectFactoryBindingEnvironment(reader, bindingAdd);
-                    break;
-                }
-                default: {
-                    throw unexpectedElement(reader);
-                }
-            }
-        }
-
-        operations.add(bindingAdd);
-    }
-
-    /**
-     * <p>
-     * Parses the optional {@code ObjectFactory environment}.
-     * </p>
-     *
-     * @param reader     the {@code XMLExtendedStreamReader} used to read the configuration XML.
-     * @param bindingAdd where to add
-     * @throws javax.xml.stream.XMLStreamException
-     *          if an error occurs while parsing the XML.
-     */
-    private void parseObjectFactoryBindingEnvironment(XMLExtendedStreamReader reader, ModelNode bindingAdd) throws XMLStreamException {
-        // no attributes expected
-        requireNoAttributes(reader);
-        while (reader.hasNext() && reader.nextTag() != XMLStreamConstants.END_ELEMENT) {
-            switch (NamingSubsystemXMLElement.forName(reader.getLocalName())) {
-                case ENVIRONMENT_PROPERTY: {
-                    final String[] array = requireAttributes(reader, org.jboss.as.controller.parsing.Attribute.NAME.getLocalName(), org.jboss.as.controller.parsing.Attribute.VALUE.getLocalName());
-                    NamingBindingResourceDefinition.ENVIRONMENT.parseAndAddParameterElement(array[0], array[1], bindingAdd, reader);
-                    requireNoContent(reader);
-                    break;
-                }
-                default: {
-                    throw unexpectedElement(reader);
-                }
-            }
-        }
-    }
-
-    private void parseLookupBinding(final XMLExtendedStreamReader reader, List<ModelNode> operations, PathAddress parentAddress) throws XMLStreamException {
-        final int attCount = reader.getAttributeCount();
-        String name = null;
-        final ModelNode bindingAdd = Util.createAddOperation();
-        bindingAdd.get(BINDING_TYPE).set(LOOKUP);
-        final EnumSet<NamingSubsystemXMLAttribute> required = EnumSet.of(NamingSubsystemXMLAttribute.NAME, NamingSubsystemXMLAttribute.LOOKUP);
-        for (int i = 0; i < attCount; i++) {
-            requireNoNamespaceAttribute(reader, i);
-            final String value = reader.getAttributeValue(i);
-            final NamingSubsystemXMLAttribute attribute = NamingSubsystemXMLAttribute.forName(reader.getAttributeLocalName(i));
-            required.remove(attribute);
-            switch (attribute) {
-                case NAME:
-                    name = value.trim();
-                    break;
-                case LOOKUP:
-                    NamingBindingResourceDefinition.LOOKUP.parseAndSetParameter(value, bindingAdd, reader);
-                    break;
-                default:
-                    throw unexpectedAttribute(reader, i);
-            }
-        }
-        if (!required.isEmpty()) {
-            throw missingRequired(reader, required);
-        }
-
-        requireNoContent(reader);
-        bindingAdd.get(OP_ADDR).set(parentAddress.append(BINDING, name).toModelNode());
-        operations.add(bindingAdd);
+        super(NamingSubsystemNamespace.NAMING_2_0);
     }
 }
