+++ naming/src/main/java/org/jboss/as/naming/subsystem/NamingSubsystemNamespace.java
@@ -38,6 +38,7 @@ public enum NamingSubsystemNamespace {
     NAMING_1_1("urn:jboss:domain:naming:1.1"),
     NAMING_1_2("urn:jboss:domain:naming:1.2"),
     NAMING_1_3("urn:jboss:domain:naming:1.3"),
+    NAMING_1_4("urn:jboss:domain:naming:1.4"),
     NAMING_2_0("urn:jboss:domain:naming:2.0"),
     ;
